pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syslang"
version = "1.0.0"
description = "DSL for transdisciplinary systemic analysis"
readme = "README.md"
authors = [
    {name = "Echopraxium", email = "echopraxium@github.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
syslang = "syslang.cli:main"

[tool.setuptools.packages.find]
where = ["src"]